requirements.txt
asgiref==3.8.1
Django==5.0.3
django-treebeard==4.7.1
djangorestframework==3.15.1
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
